Finished In-game building menu
Added transparent building display (on hover over valid tiles) when in
build mode, hotkeys (num_01-04) for building buttons, shift to keep
placing buildings, right click to cancel build mode, spend money or
cancel build mode if cannot afford the building. Added input masks to
prevent click-thrus from menus. Cleaned InputMsg ints.

+++ SpacePirates/Assets/Scripts/Display/CameraPan_hori.cs
@@ -23,7 +23,7 @@ public class CameraPan_hori : MonoBehaviour
 	private float time_remaining = 0.0f;
 	private float time_totalscrolltime = 0.15f;
 
-	private int myMsgType = InputMsg.key_horizontal;
+	private int myMsgType = (int)InputMsg.key_horizontal;
 
 	void Awake ()
 	{
